Added the first start page
Made one for a guy on a forum. He wanted something minimalistic and clean, with not many links or distractions.
His site is reachcinema.com so go check him out.

Based off http://startpages.github.io/startpages/HorizontalList-0XE4/ which is a little buggy if you're on any resolution greater than 1920x1080, so I made some small edits to it.